接口实现

+++ src/main/java/com/wwinfo/service/impl/OrganizeServiceImpl.java
@@ -7,12 +7,16 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.wwinfo.common.ExcludeEmptyQueryWrapper;
 import com.wwinfo.common.exception.BusinessException;
+import com.wwinfo.mapper.AssetMapper;
+import com.wwinfo.mapper.InvetorytaskMapper;
 import com.wwinfo.mapper.UserMapper;
+import com.wwinfo.model.Invetorytask;
 import com.wwinfo.model.Organize;
 import com.wwinfo.mapper.OrganizeMapper;
 import com.wwinfo.model.User;
 import com.wwinfo.pojo.bo.OrganizeNode;
 import com.wwinfo.pojo.query.OrganizeQuery;
+import com.wwinfo.pojo.res.AssetApiRes;
 import com.wwinfo.pojo.vo.OrganizeAddVO;
 import com.wwinfo.pojo.vo.OrganizeChgVO;
 import com.wwinfo.service.OrganizeService;
@@ -22,6 +26,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -41,6 +46,12 @@ public class OrganizeServiceImpl extends ServiceImpl<OrganizeMapper, Organize> i
 
     @Resource
     private UserMapper userMapper;
+
+    @Resource
+    private AssetMapper assetMapper;
+
+    @Resource
+    private InvetorytaskMapper invetorytaskMapper;
 
 
     @Override
@@ -62,7 +73,7 @@ public class OrganizeServiceImpl extends ServiceImpl<OrganizeMapper, Organize> i
     public IPage nextLevel(Long orgID, Integer pageSize, Integer pageNum) {
         Page<Organize> page = new Page<>(pageNum, pageSize);
         QueryWrapper<Organize> wrapper = new ExcludeEmptyQueryWrapper<>();
-        wrapper.eq("orgID", orgID);
+        wrapper.eq("upOrgID", orgID);
         return organizeMapper.selectPage(page, wrapper);
     }
 
@@ -82,7 +93,7 @@ public class OrganizeServiceImpl extends ServiceImpl<OrganizeMapper, Organize> i
         //校验orgName唯一性
         Organize exist = getOrganizeByOrgName(organizeAddVO.getOrgName());
         if(exist != null){
-            throw new BusinessException("部门名称不能为空");
+            throw new BusinessException("部门名称不能重复");
         }
         Organize organize = BeanUtil.copyProperties(organizeAddVO, Organize.class);
         return organizeMapper.insert(organize);
@@ -94,7 +105,7 @@ public class OrganizeServiceImpl extends ServiceImpl<OrganizeMapper, Organize> i
         //校验orgName唯一性
         Organize exist = getOrganizeByOrgName(organizeChgVO.getOrgName());
         if(exist != null && exist.getOrgID() != organizeChgVO.getOrgID()){
-            throw new BusinessException("部门名称不能为空");
+            throw new BusinessException("部门名称不能重复");
         }
         Organize organize = BeanUtil.copyProperties(organizeChgVO, Organize.class);
         return organizeMapper.updateById(organize);
@@ -115,8 +126,23 @@ public class OrganizeServiceImpl extends ServiceImpl<OrganizeMapper, Organize> i
         QueryWrapper<User> wrapper = new ExcludeEmptyQueryWrapper<>();
         wrapper.eq("orgID", orgID);
         List<User> userList = userMapper.selectList(wrapper);
-        if(CollUtil.isNotEmpty(userList))
+        if(CollUtil.isNotEmpty(userList)) {
             throw new BusinessException("该部门下存在用户，不能删除");
+        }
+
+        //校验是否存在资产
+        List<Long> orgList = new ArrayList<>();
+        orgList.add(orgID);
+        List<AssetApiRes> assetList = assetMapper.getAssetListByOrgs(orgList);
+        if(CollUtil.isNotEmpty(assetList)){
+            throw new BusinessException("该部门下存在资产，不能删除");
+        }
+
+        //校验是否存在任务
+        Invetorytask task = invetorytaskMapper.getTaskByOrgId(orgID);
+        if(task != null){
+            throw new BusinessException("该部门下存在盘点任务，不能删除");
+        }
         return organizeMapper.deleteById(orgID);
     }
